Improve asynchronous exceptions handling.
The problem is easily demonstrated with the following short program:

```Haskell
{-# LANGUAGE QuasiQuotes #-}

import System.Timeout (timeout)
import System.Command.QQ (sh)

main :: IO ()
main =
  print =<< timeout (1 * 1000000) (do (_, _, _) <- [sh| sleep 2; echo hello >/dev/tty|]; return ())
```

(Don't forget to use the "-threaded" runtime when compiling it.)

The expected output of the program is the string "Nothing\n".

```
% ./Main
Nothing
%
```

What actually happens is that you get your `Nothing` and then, after a
second, "hello" is printed to the console:

```
% ./Main
Nothing
% hello
```

The patch attempts to fix this. (`Eval (IO ())` is still borked, through.)